Added confirmation dialog boxes for all delete and cancel buttons.

+++ src/controller/MainMenuController.java
@@ -8,10 +8,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 import model.Inventory;
@@ -97,7 +94,16 @@ public class MainMenuController implements Initializable {
 
     @FXML
     void onActionPartsDelete(ActionEvent event) {
-        Inventory.deletePart(partsTableView.getSelectionModel().getSelectedItem());
+
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirm");
+        alert.setContentText("Are you sure you want to delete?");
+
+        alert.showAndWait();
+
+        if(alert.getResult().getButtonData() == ButtonBar.ButtonData.OK_DONE) {
+            Inventory.deletePart(partsTableView.getSelectionModel().getSelectedItem());
+        }
     }
 
     @FXML
@@ -159,7 +165,16 @@ public class MainMenuController implements Initializable {
 
     @FXML
     void onActionProductsDelete(ActionEvent event) {
-        Inventory.deleteProduct(productsTableView.getSelectionModel().getSelectedItem());
+
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirm");
+        alert.setContentText("Are you sure you want to delete?");
+
+        alert.showAndWait();
+
+        if(alert.getResult().getButtonData() == ButtonBar.ButtonData.OK_DONE) {
+            Inventory.deleteProduct(productsTableView.getSelectionModel().getSelectedItem());
+        }
     }
 
     @FXML
